parseInt was replace by parseFloat

+++ node-require/calculate.js
@@ -6,11 +6,11 @@ const divide = require('./divide');
 
 
 if (process.argv[3] === 'plus') {
-  console.log ('result: ', add(parseInt(process.argv[2]), parseInt(process.argv[4])));
+  console.log('result: ', add(parseFloat(process.argv[2]), parseFloat(process.argv[4])));
 } else if (process.argv[3] === 'minus') {
-  console.log('result: ', subtract(parseInt(process.argv[2]), parseInt(process.argv[4])));
+  console.log('result: ', subtract(parseFloat(process.argv[2]), parseFloat(process.argv[4])));
 } else if (process.argv[3] === 'times') {
-  console.log('result: ', multiply(parseInt(process.argv[2]), parseInt(process.argv[4])));
+  console.log('result: ', multiply(parseFloat(process.argv[2]), parseFloat(process.argv[4])));
 } else if (process.argv[3] === 'over') {
-  console.log('result: ', divide(parseInt(process.argv[2]), parseInt(process.argv[4])));
+  console.log('result: ', divide(parseFloat(process.argv[2]), parseFloat(process.argv[4])));
 }
